# src/main/resources/mastercard-send-customer-parameter-management-api-swagger.yaml
openapi: 3.0.3
info:
  title: Mastercard Send Customer Parameter Management API
  description: >-
    This API is for Sponsor Banks (Originating Institutions) who are working
    with partners (Transaction Initiators) to provide funds transfer solutions
    (programs) using Mastercard Send. You can use this API to set limits for
    your partners and their Transfer Acceptors, configure alerts, and retrieve
    the current daily usage figures.

    Types of configurations:

    * Partner limit configuration = Daily limits and alert settings for a
      partner

    * Transfer Acceptor limit configuration = Daily limits and alert settings
      for a Transfer Acceptor

    * Account limit configuration = Account-level settings and limits
      for a partner, which can apply to either a specific Transfer Acceptor's
      transactions or all that partner's transactions

    NOTE: For the `partner_id` portion of the URL, use the Mastercard-assigned
    unique Partner Reference ID for the partner (Transaction Initiator) whose
    configuration you are viewing or setting. Do not use your Partner Reference
    ID.
  version: 1.0.0
  contact:
    name: API Support
    url: 'https://developer.mastercard.com/support'
    email: apisupport@mastercard.com
servers:
  - url: 'https://sandbox.api.mastercard.com/send/partners'
    description: Sandbox URL
  - url: 'https://mtf.api.mastercard.com/send/partners'
    description: MTF URL
  - url: 'https://api.mastercard.com/send/partners'
...
paths:
...
  '/{partner_id}/limits/daily-limits/transfer-acceptor-limits':
    post:
      operationId: postTransferAcceptorLimit
      summary: Add daily limits and alert settings for a Transfer Acceptor.
      description: >-
        Adds a Transfer Acceptor limit configuration for the partner identified
        in the path (`partner_id`). The success (201) response will include the
        reference ID for the Transfer Acceptor limit configuration
        (`transferAcceptorLimitReference`), which you can use to update the
        configuration.
      tags:
        - Transfer Acceptor Limits
      parameters:
        - $ref: '#/components/parameters/partner_id'
        - $ref: '#/components/parameters/Accept'
        - $ref: '#/components/parameters/Content-Type'
      requestBody:
        $ref: '#/components/requestBodies/PostTransferAcceptorLimit'
      responses:
        '201':
          $ref: '#/components/responses/PostTransferAcceptorLimitResponse'
        '400':
          $ref: '#/components/responses/Error400Response'
        '401':
          $ref: '#/components/responses/Error401Response'
        '409':
          $ref: '#/components/responses/TransferAcceptorError409Response'
  '/{partner_id}/limits/transfer-acceptor-limits/{transfer_acceptor_limit_reference}':
    put:
      operationId: putTransferAcceptorLimit
      summary: Update daily limits and alert settings for a Transfer Acceptor.
      description: >-
        Updates the Transfer Acceptor limit configuration identified in the path
        (`transfer_acceptor_limit_reference`). Your request must include the
        `dailyLimit` object, but the `alertSettings` object is optional. If you
        omit the `alertSettings` object, that object's previous settings will be
        retained.
      tags:
        - Transfer Acceptor Limits
      parameters:
        - $ref: '#/components/parameters/partner_id'
        - $ref: '#/components/parameters/transfer_acceptor_limit_reference'
        - $ref: '#/components/parameters/Accept'
        - $ref: '#/components/parameters/Content-Type'
      requestBody:
        $ref: '#/components/requestBodies/PutTransferAcceptorLimit'
      responses:
        '200':
          $ref: '#/components/responses/PutTransferAcceptorLimitResponse'
        '400':
          $ref: '#/components/responses/Error400Response'
        '401':
          $ref: '#/components/responses/Error401Response2'
        '404':
          $ref: '#/components/responses/TransferAcceptorError404Response'
...
components:
  responses:
...
    PostTransferAcceptorLimitResponse:
      description: >-
        CREATED. The Transfer Acceptor limit configuration has been added for
        the Transfer Acceptor. Returns the new configuration details, including
        its reference ID (`transferAcceptorLimitReference`).
      headers:
        Correlation-ID:
          $ref: '#/components/headers/Correlation-ID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TransferAcceptorLimitParent'
    PutTransferAcceptorLimitResponse:
      description: >-
        OK. The Transfer Acceptor limit configuration has been updated for the
        Transfer Acceptor. Returns the updated configuration details.
      headers:
        Correlation-ID:
          $ref: '#/components/headers/Correlation-ID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TransferAcceptorLimitParent'
...
    Error400Response:
      description: BAD REQUEST. Returns an array of errors when request fields are invalid.
      headers:
        Correlation-ID:
          $ref: '#/components/headers/Correlation-ID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            ErrorResponse:
              $ref: '#/components/examples/ErrorResponse400'
...
    Error401Response:
      description: >-
        UNAUTHORIZED. Returns an array of errors in case of missing or invalid
        authentication token.
      headers:
        Correlation-ID:
          $ref: '#/components/headers/Correlation-ID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            ErrorResponse1:
              $ref: '#/components/examples/ErrorResponse401Unauthorized'
            ErrorResponse2:
              $ref: '#/components/examples/ErrorResponse401'
    Error401Response2:
      description: >-
        UNAUTHORIZED. Returns an array of errors in case of missing or invalid
        authentication token.
      headers:
        Correlation-ID:
          $ref: '#/components/headers/Correlation-ID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            ErrorResponse1:
              $ref: '#/components/examples/ErrorResponse401Unauthorized'
...
    TransferAcceptorError404Response:
      description: >-
        NOT FOUND. Returns an array of errors when no Transfer Acceptor limit
        configuration can be found for the provided
        `transfer_acceptor_limit_reference`.
      headers:
        Correlation-Id:
          $ref: '#/components/headers/Correlation-ID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            ErrorResponse:
              $ref: '#/components/examples/TransferAcceptorErrorResponse404'
...
    TransferAcceptorError409Response:
      description: CONFLICT. The Transfer Acceptor limit configuration is already present.
      headers:
        Correlation-ID:
          $ref: '#/components/headers/Correlation-ID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            ErrorResponse:
              $ref: '#/components/examples/TransferAcceptorErrorResponse409'
...
  requestBodies:
...
    PostTransferAcceptorLimit:
      description: Data for adding a Transfer Acceptor limit configuration.
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TransferAcceptorLimitPostDetailParent'
    PutTransferAcceptorLimit:
      description: Data for updating a Transfer Acceptor limit configuration.
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TransferAcceptorLimitPutDetailParent'
...
  schemas:
...
    TransferAcceptorDailyLimitParent:
      type: object
      description: Object specifying the daily limits for the Transfer Acceptor.
      properties:
        transactionAmountLimit:
          $ref: '#/components/schemas/TransferAcceptorTransactionAmountLimit'
        transactionVolumeLimit:
          $ref: '#/components/schemas/TransferAcceptorTransactionVolumeLimit'
    TransferAcceptorDailyLimitSetParent:
      type: object
      description: >-
        Object specifying the daily limits for the Transfer Acceptor. You cannot
        set the limits higher than the partner daily limits.
      required:
        - transactionAmountLimit
      properties:
        transactionAmountLimit:
          $ref: '#/components/schemas/TransferAcceptorTransactionAmountLimit'
        transactionVolumeLimit:
          $ref: '#/components/schemas/TransferAcceptorTransactionVolumeLimit'
...
    TransferAcceptorLimitPutDetailParent:
      type: object
      description: Request object for setting the Transfer Acceptor limits.
      required:
        - dailyLimit
      properties:
        dailyLimit:
          $ref: '#/components/schemas/TransferAcceptorDailyLimitSetParent'
        alertSettings:
          $ref: '#/components/schemas/TransferAcceptorAlertSettings'
    TransferAcceptorLimitPostDetailParent:
      type: object
      description: Request object for setting the Transfer Acceptor limits.
      required:
        - transferAcceptorId
        - dailyLimit
      properties:
        transferAcceptorId:
          type: string
          nullable: false
          minLength: 15
          maxLength: 15
          description: >-
            An identifier for the Transfer Acceptor. This value should be unique
            for each originator.
          example: '456487898368123'
        dailyLimit:
          $ref: '#/components/schemas/TransferAcceptorDailyLimitSetParent'
        alertSettings:
          $ref: '#/components/schemas/TransferAcceptorAlertSettings'
...
    TransferAcceptorLimitParent:
      type: object
      description: Response object for setting the Transfer Acceptor limits.
      properties:
        transferAcceptorLimitReference:
          type: string
          description: >-
            The reference ID for the Transfer Acceptor limit configuration,
            which was either assigned during the onboarding process or returned
            when the configuration was added. String, length 1-40.
          minLength: 1
          maxLength: 40
          nullable: false
          example: TA_Ref_23423
        transferAcceptorId:
          type: string
          nullable: false
          minLength: 15
          maxLength: 15
          description: >-
            An identifier for the Transfer Acceptor. This value should be unique
            for each originator.
          example: '456487898368123'
        dailyLimit:
          $ref: '#/components/schemas/TransferAcceptorDailyLimitParent'
        alertSettings:
          $ref: '#/components/schemas/TransferAcceptorAlertSettings'
...
    TransferAcceptorAlertSettings:
      type: object
      description: Object containing the alert settings for the Transfer Acceptor.
      required:
        - alertEmails
        - amountAlertThreshold
      properties:
        alertEmails:
          type: array
          description: >-
            The email address(es) of the associated entities (business partners
            or Sponsor Banks) who are entitled to receive alert notifications.
            This is required if the business partner has opted for alert
            notifications.
          items:
            type: string
            format: email
          nullable: true
          maxLength: 1500
          example:
            - test123@abcmail.com
            - test456@abcmail.com
        amountAlertThreshold:
          type: number
          description: >-
            The amount alert threshold, which is a percentage (e.g. 95). When
            the Transfer Acceptor exceeds this percentage of their daily amount
            limit (`transactionAmountLimit.amount`), alert notifications will be
            sent if alerts are enabled (`alertsEnabled` = true, in the partner
            limit configuration). Numeric integer, do not include the % symbol.
          minLength: 1
          maxLength: 3
          example: 95
        volumeAlertThreshold:
          type: number
          description: >-
            The volume alert threshold, which is a percentage (e.g. 80). When
            the Transfer Acceptor exceeds this percentage of their daily volume
            limit (`maxTransactionVolume`), alert notifications will be sent if
            alerts are enabled (`alertsEnabled` = true, in the partner limit
            configuration). Numeric integer, do not include the % symbol.
          minLength: 1
          maxLength: 3
          example: 80
...
    TransferAcceptorTransactionAmountLimit:
      type: object
      description: Daily amount limit for the Transfer Acceptor.
      required:
        - amount
        - currency
      properties:
        amount:
          type: string
          format: integer
          minLength: 1
          maxLength: 13
          description: >-
            Daily transaction limit for the Transfer Acceptor (the maximum
            amount permitted to be transacted per day), in the currency
            identified by the `currency` field. It is a positive non-zero
            integer, length 1-13. The decimal position is implied based on the
            relevant currency exponent; see [Country and Currency
            Codes](https://developer.mastercard.com/mastercard-send-2-0/documentation/field-uris-codes/country-currency-codes/).
            For example, 1000000 Indian Rupees (INR) will be a value of
            100000000.
          nullable: false
          example: '100000000'
        currency:
          type: string
          description: >-
            The 3-character ISO 4217 alpha-3 code identifying the currency for
            the amount in the `amount` field. See [Country and
            Currency Codes](https://developer.mastercard.com/mastercard-send-2-0/documentation/field-uris-codes/country-currency-codes/).
            For example, for Indian Rupees, the value is INR.
          minLength: 3
          maxLength: 3
          nullable: false
          example: INR
    TransferAcceptorTransactionVolumeLimit:
      type: object
      description: Daily volume limit for the Transfer Acceptor.
      properties:
        maxTransactionVolume:
          type: string
          format: integer
          minLength: 0
          maxLength: 8
          description: The maximum number of transactions permitted per day.
          nullable: true
          example: '50000'
    ErrorResponse:
      type: object
      description: >-
        Response object returned in case of any errors that occurred during
        processing.
      required:
        - Errors
      properties:
        Errors:
          type: object
          description: Parent object containing the Error array.
          required:
            - Error
          properties:
            Error:
              $ref: '#/components/schemas/Errors'
    Errors:
      type: array
      description: Parent object containing the array.
      minItems: 1
      items:
        $ref: '#/components/schemas/Error'
    Error:
      type: object
      description: An array of errors.
      properties:
        Source:
          type: string
          description: >-
            Where possible, identifies the field that is causing the error, for
            example an incorrectly formatted field or a missing required field.
            If the error is not associated with a specific field, system error
            is returned.
          minLength: 1
          maxLength: 100
          example: partner_id
        ReasonCode:
          type: string
          description: The general cause of the error, see [Error Codes](#error-codes).
          minLength: 1
          maxLength: 100
          example: INVALID_INPUT_VALUE
        Description:
          type: string
          description: >-
            Describes the error. Present if more information is available than
            that provided by the `ReasonCode` field.
          minLength: 1
          maxLength: 500
          example: Invalid value
        Recoverable:
          type: boolean
          description: >-
            Indicates whether the transaction can be recovered (might be
            successful) if the failed request is resubmitted. Valid values:
            true, false.
          example: false
        ErrorDetailCode:
          type: string
          description: The error code.
          minLength: 0
          maxLength: 100
          example: '082000'
        Details:
          type: string
          description: Always null, present for backwards compatibility.
          minLength: 0
          maxLength: 4
          example: 'null'
  examples:
    ErrorResponse400:
      value:
        Errors:
          Error:
            - Source: dailyLimit.transactionAmountLimit.currency
              ReasonCode: INVALID_INPUT_VALUE
              Description: Invalid value
              Recoverable: false
              ErrorDetailCode: '082000'
              Details: null
...
    ErrorResponse401:
      value:
        Errors:
          Error:
            - Source: partner_id
              ReasonCode: INVALID_INPUT_VALUE
              Description: Invalid value
              Recoverable: false
              ErrorDetailCode: '082000'
              Details: null
    ErrorResponse401Unauthorized:
      value:
        Errors:
          Error:
            - Source: partner_id
              ReasonCode: UNAUTHORIZED_ACCESS
              Description: Unauthorized Access
              Recoverable: false
              ErrorDetailCode: '050007'
              Details: null
    TransferAcceptorErrorResponse409:
      value:
        Errors:
          Error:
            - Source: transferAcceptorId
              ReasonCode: RESOURCE_ERROR
              Description: Duplicate value
              Recoverable: false
              ErrorDetailCode: '110501'
              Details: null
...
    TransferAcceptorErrorResponse404:
      value:
        Errors:
          Error:
            - Source: transferAcceptorLimitReference
              ReasonCode: INVALID_INPUT_VALUE
              Description: Invalid value
              Recoverable: false
              ErrorDetailCode: '082000'
              Details: null
...
  parameters:
    partner_id:
      in: path
      name: partner_id
      schema:
        type: string
        minLength: 32
        maxLength: 40
      required: true
      description: >-
        The Mastercard-assigned unique Partner Reference ID for the partner
        (Transaction Initiator) whose configuration you are viewing or setting.
        It is not your Partner Reference ID. String, maximum length 40.
      example: ptnr_1lzs0zD6uWo7k-OdUXjLW_pFgX2
...
    transfer_acceptor_limit_reference:
      in: path
      name: transfer_acceptor_limit_reference
      schema:
        type: string
        minLength: 1
        maxLength: 40
      required: true
      description: >-
        The reference ID for the Transfer Acceptor limit configuration, which
        was either assigned during the onboarding process or returned when the
        configuration was added. String, length 1-40.
      example: TA_Ref_23423
...
    Accept:
      in: header
      name: Accept
      schema:
        type: string
      description: >-
        This request header indicates the type of response payload expected by
        the client. API can accept only application/json as valid values.
      example: application/json
    Content-Type:
      in: header
      name: Content-Type
      schema:
        type: string
      description: >-
        This request header indicates the payload format of the request. The API
        can accept only application/json as the valid value. The same header
        will be used in response to indicate the response payload format. API
        will only generate response payload in application/json.
      example: application/json
  headers:
    Correlation-ID:
      schema:
        type: string
      description: >-
        ID for research purposes to uniquely identify the request within
        Mastercard.
      example: cor_45345lkjflkdjsflkdf